requirements.txt
asgiref==3.5.0
astroid==2.11.7
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
colorama==0.4.5
coverage==6.4.2
cryptography==37.0.2
dill==0.3.5.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
Faker==12.0.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mixer==7.2.2
numpy==1.22.4
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
paramiko==2.11.0
PGPy==0.5.4
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tzdata==2021.5
urllib3==1.26.9
wrapt==1.14.1
